+++ Reach/src/utopia/reach/window/ReachWindow.scala
@@ -277,7 +277,6 @@ case class ContextualReachWindowFactory(context: ReachWindowContext)(implicit ex
 	}
 	
 	private def revalidate(window: Window, canvas: Stackable) = {
-		println("Revalidating windows and canvas")
 		// Resets cached stack sizes in order to make sure the sizes are set correctly
 		canvas.resetCachedSize()
 		window.resetCachedSize()